// frontend/src/api-generated/models/chiTietBaoTriGeneralResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface ChiTietBaoTriGeneralResponse {
  id?: number;
  idTaiSanGoc?: number;
  tenMauTaiSan?: string;
  soSerial?: string;
  maTheTaiSan?: string;
  loaiHinhXuLy?: string;
  idNhaCungCap?: number;
  tenNhaCungCap?: string;
  trangThaiThucHien?: string;
  tinhTrangThietBi?: string;
  phuongAnXuLy?: string;
  idLinhKienThayThe?: number;
  chiPhi?: number;
  loai?: string;
}